修改UserModel解析，double时报错

+++ lib/common/model/user/user_model.dart
@@ -58,7 +58,7 @@ class UserInfo {
   int? agentLevel;
   int? countPool;
   int? childs;
-  double? product;
+  int? product;
   int? teamTotal;
   TjrInfo? tjrInfo;
   String? avatars;
